src/main/java/com/factorit/carritocompras/service/UsuarioService.java
package com.factorit.carritocompras.service;

import com.factorit.carritocompras.domain.Usuario;
import com.factorit.carritocompras.repository.UsuarioRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class UsuarioService {

    @Autowired
    private UsuarioRepository usuarioRepository;

    public List<Usuario> buscarTodos(){
        return usuarioRepository.findAll();
    }

    public Usuario buscarPorId(int id) {
        return usuarioRepository.findById(id).orElse(null);
    }
}

